requirements.txt
krakenex==2.1.0
pykrakenapi==0.3.2
groq==0.4.1
anthropic==0.7.7
ollama==0.1.7
python-dotenv==1.0.0
numpy==1.24.3
scipy==1.10.1
arch==5.3.1
hmmlearn==0.3.0
PyWavelets==1.4.1
libcst==1.0.1
websockets==11.0.3
asyncio-mqtt==0.13.0